meta sprite atlas saves to disc

+++ TileMapRendererExperiments/BasicTypedefs.h
@@ -6,7 +6,7 @@ typedef uint8_t u8;
 
 typedef int8_t i8;
 typedef int16_t i16;
-typedef int16_t i32;
+typedef int32_t i32;
 
 typedef float f32;
 typedef double f64;
